Updating to include sup application disabled button logic.
