+++ frontend/src/services/applicationService.js
@@ -1,8 +1,10 @@
-import { ApiRoutes } from '@/utils/constants'
-import ApiService from '@/common/apiService'
 import { isEmpty } from 'lodash'
+
+import ApiService from '@/common/apiService'
 import { useApplicationsStore } from '@/stores/applications'
 import { useAuthStore } from '@/stores/auth'
+import { ApiRoutes } from '@/utils/constants'
+import { APPLICATION_STATUS_CODES, CRM_STATE_CODES } from '@/utils/constants'
 
 function sortApplications(applications) {
   applications?.sort((app1, app2) => {
@@ -122,5 +124,25 @@ export default {
       console.log(`Failed to get the supp application by application id - ${error}`)
       throw error
     }
+  },
+
+  isValidApplication(application) {
+    return this.checkApplicationStatus(application) && this.checkFundingAgreement(application?.fundingAgreements)
+  },
+
+  checkApplicationStatus(application) {
+    const isActive = application?.stateCode === CRM_STATE_CODES.ACTIVE
+    const hasCorrectStatus = [APPLICATION_STATUS_CODES.SUBMITTED, APPLICATION_STATUS_CODES.IN_REVIEW, APPLICATION_STATUS_CODES.APPROVED, APPLICATION_STATUS_CODES.AWAITING_PROVIDER].includes(
+      application?.statusCode,
+    )
+    return isActive && hasCorrectStatus
+  },
+
+  checkFundingAgreement(fundingAgreements) {
+    return fundingAgreements?.some((funding) => {
+      const isActive = funding?.stateCode === CRM_STATE_CODES.ACTIVE
+      const isUnexpired = new Date() < new Date(funding?.endDate)
+      return isActive && isUnexpired
+    })
   },
 }
