+++ frontend/src/utils/constants.js
@@ -62,13 +62,6 @@ export const APPLICATION_STATUS_CODES = Object.freeze({
   APPROVED: 6,
 })
 
-export const APPLICATION_ACTIVE_STATUS_CODES = Object.freeze([
-  APPLICATION_STATUS_CODES.SUBMITTED,
-  APPLICATION_STATUS_CODES.IN_REVIEW,
-  APPLICATION_STATUS_CODES.AWAITING_PROVIDER,
-  APPLICATION_STATUS_CODES.APPROVED,
-])
-
 export const FACILITY_TYPES = Object.freeze({
   RENT_LEASE: 1,
   OWNED_WITH_MORTGAGE: 2,
